+++ src/controllers/box/box.controller.ts
@@ -7,13 +7,24 @@ import ResponseHandler from '../../utils/responsesHandler';
 import SystemLogModel from '../../models/logs/system.log.model';
 import authHandler from '../../utils/authHandler';
 import AuditTrailModel from '../../models/logs/audit.trail.model';
+import CountryModel from '../../models/adminstration/country.model';
+import CityModel from '../../models/adminstration/city.model';
+import UserBoxModel from '../../models/box/user.box.model';
+import TabletModel from '../../models/box/tablet.model';
 
+const userBoxModel = new UserBoxModel();
+const countryModel = new CountryModel();
+const cityModel = new CityModel();
 const auditTrail = new AuditTrailModel();
 const systemLog = new SystemLogModel();
 const boxModel = new BoxModel();
+const tabletModel = new TabletModel();
 
 export const createBox = asyncHandler(async (req: Request, res: Response) => {
   const user = await authHandler(req, res);
+  if (user === '0') {
+    return user;
+  }
 
   try {
     const newBox: Box = req.body;
@@ -25,17 +36,18 @@ export const createBox = asyncHandler(async (req: Request, res: Response) => {
     }
 
     const createdBox = await boxModel.createBox(newBox);
-    ResponseHandler.success(
-      res,
-      i18n.__('BOX_CREATED_SUCCESSFULLY'),
-      createdBox,
-    );
+
     const action = 'createBox';
     auditTrail.createAuditTrail(
       user,
       action,
       i18n.__('BOX_CREATED_SUCCESSFULLY'),
       createdBox.id,
+    );
+    ResponseHandler.success(
+      res,
+      i18n.__('BOX_CREATED_SUCCESSFULLY'),
+      createdBox,
     );
   } catch (error) {
     const source = 'createBox';
@@ -55,6 +67,9 @@ export const getAllBoxes = asyncHandler(async (req: Request, res: Response) => {
     );
   } catch (error) {
     const user = await authHandler(req, res);
+    if (user === '0') {
+      return user;
+    }
     const source = 'getAllBoxes';
     systemLog.createSystemLog(user, (error as Error).message, source);
     ResponseHandler.badRequest(res, (error as Error).message);
@@ -64,6 +79,9 @@ export const getAllBoxes = asyncHandler(async (req: Request, res: Response) => {
 
 export const getBoxById = asyncHandler(async (req: Request, res: Response) => {
   const user = await authHandler(req, res);
+  if (user === '0') {
+    return user;
+  }
   try {
     const boxId = req.params.id;
     const box = await boxModel.getOne(boxId);
@@ -78,22 +96,26 @@ export const getBoxById = asyncHandler(async (req: Request, res: Response) => {
 
 export const updateBox = asyncHandler(async (req: Request, res: Response) => {
   const user = await authHandler(req, res);
+  if (user === '0') {
+    return user;
+  }
 
   try {
     const boxId = req.params.id;
     const boxData: Partial<Box> = req.body;
     const updatedBox = await boxModel.updateOne(boxData, boxId);
-    ResponseHandler.success(
-      res,
-      i18n.__('BOX_UPDATED_SUCCESSFULLY'),
-      updatedBox,
-    );
+
     const action = 'updateBox';
     auditTrail.createAuditTrail(
       user,
       action,
       i18n.__('BOX_UPDATED_SUCCESSFULLY'),
       updatedBox.id,
+    );
+    ResponseHandler.success(
+      res,
+      i18n.__('BOX_UPDATED_SUCCESSFULLY'),
+      updatedBox,
     );
   } catch (error) {
     const source = 'updateBox';
@@ -105,21 +127,25 @@ export const updateBox = asyncHandler(async (req: Request, res: Response) => {
 
 export const deleteBox = asyncHandler(async (req: Request, res: Response) => {
   const user = await authHandler(req, res);
+  if (user === '0') {
+    return user;
+  }
 
   try {
     const boxId = req.params.id;
     const deletedBox = await boxModel.deleteOne(boxId);
-    ResponseHandler.success(
-      res,
-      i18n.__('BOX_DELETED_SUCCESSFULLY'),
-      deletedBox,
-    );
+
     const action = 'deleteBox';
     auditTrail.createAuditTrail(
       user,
       action,
       i18n.__('BOX_DELETED_SUCCESSFULLY'),
       deletedBox.id,
+    );
+    ResponseHandler.success(
+      res,
+      i18n.__('BOX_DELETED_SUCCESSFULLY'),
+      deletedBox,
     );
   } catch (error) {
     const source = 'deleteBox';
@@ -132,6 +158,9 @@ export const deleteBox = asyncHandler(async (req: Request, res: Response) => {
 export const getBoxesByGenerationId = asyncHandler(
   async (req: Request, res: Response) => {
     const user = await authHandler(req, res);
+    if (user === '0') {
+      return user;
+    }
 
     try {
       const boxGenerationId = req.params.generationId;
@@ -154,6 +183,9 @@ export const getBoxesByGenerationId = asyncHandler(
 export const getBoxByTabletInfo = asyncHandler(
   async (req: Request, res: Response) => {
     const user = await authHandler(req, res);
+    if (user === '0') {
+      return user;
+    }
 
     try {
       const { androidTabletId, tabletSerialNumber } = req.body;
@@ -174,6 +206,9 @@ export const getBoxByTabletInfo = asyncHandler(
 export const assignTabletToBox = asyncHandler(
   async (req: Request, res: Response) => {
     const user = await authHandler(req, res);
+    if (user === '0') {
+      return user;
+    }
 
     try {
       const { tabletId, boxId } = req.body;
@@ -181,17 +216,18 @@ export const assignTabletToBox = asyncHandler(
         tabletId,
         boxId,
       );
-      ResponseHandler.success(
-        res,
-        i18n.__('TABLET_ASSIGNED_TO_BOX_SUCCESSFULLY'),
-        assignTabletToBox,
-      );
+
       const action = 'assignTabletToBox';
       auditTrail.createAuditTrail(
         user,
         action,
         i18n.__('TABLET_ASSIGNED_TO_BOX_SUCCESSFULLY'),
         boxId,
+      );
+      ResponseHandler.success(
+        res,
+        i18n.__('TABLET_ASSIGNED_TO_BOX_SUCCESSFULLY'),
+        assignTabletToBox,
       );
     } catch (error) {
       const source = 'assignTabletToBox';
@@ -205,14 +241,22 @@ export const assignTabletToBox = asyncHandler(
 export const resetTabletId = asyncHandler(
   async (req: Request, res: Response) => {
     const user = await authHandler(req, res);
+    if (user === '0') {
+      return user;
+    }
     try {
       const { tabletId, boxId } = req.body;
+
+      // check if the tablet alraedy has a box
+      const tablet = await tabletModel.tabletIsAssignedToBox(tabletId);
+      if (tablet) {
+        return ResponseHandler.badRequest(
+          res,
+          i18n.__('TABLET_ALREADY_ASSIGNED_TO_BOX'),
+        );
+      }
       const resetedTablte = await boxModel.resetTabletId(tabletId, boxId);
-      ResponseHandler.success(
-        res,
-        i18n.__('TABLET_RESET_TO_BOX_SUCCESSFULLY'),
-        resetedTablte,
-      );
+
       const action = 'resetTabletId';
       auditTrail.createAuditTrail(
         user,
@@ -220,8 +264,77 @@ export const resetTabletId = asyncHandler(
         i18n.__('TABLET_RESET_TO_BOX_SUCCESSFULLY'),
         boxId,
       );
+      ResponseHandler.success(
+        res,
+        i18n.__('TABLET_RESET_TO_BOX_SUCCESSFULLY'),
+        resetedTablte,
+      );
     } catch (error) {
       const source = 'resetTabletId';
+      systemLog.createSystemLog(user, (error as Error).message, source);
+      ResponseHandler.badRequest(res, (error as Error).message);
+      // next(error);
+    }
+  },
+);
+
+export const updateBoxAndAddress = asyncHandler(
+  async (req: Request, res: Response) => {
+    const user = await authHandler(req, res);
+    if (user === '0') {
+      return user;
+    }
+
+    try {
+      const boxId = req.params.id;
+      const { boxLabel, country_id, city_id, district, street } = req.body;
+
+      if (!boxId) {
+        return ResponseHandler.badRequest(res, i18n.__('BOX_ID_REQUIRED'));
+      }
+
+      const boxRelatedToUser = await userBoxModel.checkUserBox(user, boxId);
+
+      if (!boxRelatedToUser) {
+        return ResponseHandler.badRequest(
+          res,
+          i18n.__('BOX_NOT_RELATED_TO_USER'),
+        );
+      }
+
+      // check if the city and country exist
+      const countryExist = await countryModel.getOne(country_id);
+      if (!countryExist) {
+        return ResponseHandler.badRequest(res, i18n.__('COUNTRY_NOT_EXIST'));
+      }
+
+      const cityExist = await cityModel.getCityById(city_id);
+      if (!cityExist) {
+        return ResponseHandler.badRequest(res, i18n.__('CITY_NOT_EXIST'));
+      }
+      const updatedBox = await boxModel.updateBoxAndAddress(
+        boxId,
+        boxLabel,
+        country_id,
+        city_id,
+        district,
+        street,
+      );
+
+      const action = 'updateBoxAndAddress';
+      auditTrail.createAuditTrail(
+        user,
+        action,
+        i18n.__('BOX_UPDATED_SUCCESSFULLY'),
+        null,
+      );
+      ResponseHandler.success(
+        res,
+        i18n.__('BOX_UPDATED_SUCCESSFULLY'),
+        updatedBox,
+      );
+    } catch (error) {
+      const source = 'updateBoxAndAddress';
       systemLog.createSystemLog(user, (error as Error).message, source);
       ResponseHandler.badRequest(res, (error as Error).message);
       // next(error);
